structured properties patch refactor
